fix: 카테고리별로 namespace를 정리함
