+++ Sources/Montage/Components/Input/InputLabelController.swift
@@ -9,8 +9,8 @@ import SwiftUI
 
 /// ``Montage/InputLabel``을 SwiftUI에서 사용할 수 있도록 감싼 컨테이너 객체입니다.
 struct InputLabelController: UIViewRepresentable {
-    @State public var inputView: MontageInput
-    @State public var state: MontageInputState
+    @State public var inputView: MontageControl
+    @State public var state: MontageControlState
     @State public var text: String
 
     public typealias UIViewType = InputLabel
@@ -28,8 +28,8 @@ struct InputLabelController: UIViewRepresentable {
 struct MontageInputLabelController_Previews: PreviewProvider {
     static var previews: some View {
         VStack {
-            InputLabelController(inputView: Radio(), state: .checked, text: "체크해주세요!").fixedSize()
-            InputLabelController(inputView: Radio(), state: .unchecked, text: "체크해주세요!").fixedSize()
+            InputLabelController(inputView: Control.Radio(), state: .checked, text: "체크해주세요!").fixedSize()
+            InputLabelController(inputView: Control.Radio(), state: .unchecked, text: "체크해주세요!").fixedSize()
         }
     }
 }
