+++ Sources/Montage/Element/Basic/Basic.swift
@@ -1,0 +1,10 @@
+//
+//  Basic.swift
+//  Montage
+//
+//  Created by Euigyom Kim on 2023/04/07.
+//
+
+import Foundation
+
+public enum Basic {}
