+++ Sources/Montage/Element/Control/Control.swift
@@ -1,0 +1,10 @@
+//
+//  Control.swift
+//  Montage
+//
+//  Created by Euigyom Kim on 2023/04/07.
+//
+
+import Foundation
+
+public enum Control {}
