+++ Sources/Montage/Element/Control/Radio/Radio.swift
@@ -11,70 +11,72 @@ import UIKit
 public protocol RadioControlDelegate: AnyObject {
     /// 터치가 발생하였을 때 호출되는 메소드입니다.
     /// - Parameter radio: 터치가 발생한 객체
-    func didTappedRadio(_ radio: Radio)
+    func didTappedRadio(_ radio: Control.Radio)
 }
 
-/// 원형 안의 점 모양을 표시하는 Control Element 입니다. ``MontageInputState``의 일부만을 표현할 수 있습니다.
-public final class Radio: UIView, MontageInput {
-    private enum Const {
-        static let wrapperBoxSize: CGSize = .init(width: 24, height: 24)
-        static let boxInset: CGFloat = .spacing(.pt02)
-    }
-    
-    /// Control Element의 모양을 표현하기 위한 상태값입니다.
-    /// `.partial` 상태는 지원하지 않으며, 해당 상태로 설정할 경우 자동으로 `.checked` 상태로 변경후 설정됩니다.
-    public var state: MontageInputState = .unchecked {
-        didSet {
-            if state == .partial {
-                #warning("Can't use .partial type on Radio element. Must use .checked type.")
-                state = .checked
+extension Control {
+    /// 원형 안의 점 모양을 표시하는 Control Element 입니다. ``MontageControl``의 일부만을 표현할 수 있습니다.
+    public final class Radio: UIView, MontageControl {
+        private enum Const {
+            static let wrapperBoxSize: CGSize = .init(width: 24, height: 24)
+            static let boxInset: CGFloat = .spacing(.pt02)
+        }
+        
+        /// Control Element의 모양을 표현하기 위한 상태값입니다.
+        /// `.partial` 상태는 지원하지 않으며, 해당 상태로 설정할 경우 자동으로 `.checked` 상태로 변경후 설정됩니다.
+        public var state: MontageControlState = .unchecked {
+            didSet {
+                if state == .partial {
+                    #warning("Can't use .partial type on Radio element. Must use .checked type.")
+                    state = .checked
+                }
+                updateViews()
             }
-            updateViews()
         }
-    }
-    
-    private lazy var boxView = UIView()
-    
-    private lazy var imageView: UIImageView = {
-        let view = UIImageView()
-        view.isUserInteractionEnabled = false
-        view.tintColor = .alias(.staticWhite)
-        return view
-    }()
-    
-    private var tapRecognizer: UITapGestureRecognizer?
-    
-    private weak var delegate: RadioControlDelegate?
-    
-    /// Radio 객체를 생성합니다.
-    /// - Parameter delegate: Radio 버튼을 Element 단독으로 사용할 경우 이벤트를 받을 delegate 객체입니다.
-    public init(delegate: RadioControlDelegate? = nil) {
-        super.init(frame: .zero)
-        self.delegate = delegate
         
-        setupViews()
-        bindEvent()
-    }
-    
-    override public required init?(coder: NSCoder) {
-        super.init(coder: coder)
+        private lazy var boxView = UIView()
         
-        setupViews()
-    }
-    
-    override public func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
-        if traitCollection.hasDifferentColorAppearance(comparedTo: previousTraitCollection) {
-            updateViews()
+        private lazy var imageView: UIImageView = {
+            let view = UIImageView()
+            view.isUserInteractionEnabled = false
+            view.tintColor = .alias(.staticWhite)
+            return view
+        }()
+        
+        private var tapRecognizer: UITapGestureRecognizer?
+        
+        private weak var delegate: RadioControlDelegate?
+        
+        /// Radio 객체를 생성합니다.
+        /// - Parameter delegate: Radio 버튼을 Element 단독으로 사용할 경우 이벤트를 받을 delegate 객체입니다.
+        public init(delegate: RadioControlDelegate? = nil) {
+            super.init(frame: .zero)
+            self.delegate = delegate
+            
+            setupViews()
+            bindEvent()
         }
-    }
-    
-    /// Element의 기본적인 사이즈를 정의합니다.
-    override public var intrinsicContentSize: CGSize {
-        Const.wrapperBoxSize
+        
+        override public required init?(coder: NSCoder) {
+            super.init(coder: coder)
+            
+            setupViews()
+        }
+        
+        override public func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+            if traitCollection.hasDifferentColorAppearance(comparedTo: previousTraitCollection) {
+                updateViews()
+            }
+        }
+        
+        /// Element의 기본적인 사이즈를 정의합니다.
+        override public var intrinsicContentSize: CGSize {
+            Const.wrapperBoxSize
+        }
     }
 }
 
-extension Radio {
+extension Control.Radio {
     private func setupViews() {
         addSubview(boxView)
         boxView.addSubview(imageView)
