+++ Sources/Montage/Element/Decorate/Decorate.swift
@@ -1,0 +1,10 @@
+//
+//  Decorate.swift
+//  Montage
+//
+//  Created by Euigyom Kim on 2023/04/07.
+//
+
+import Foundation
+
+public enum Decorate {}
