+++ Sources/Montage/Element/Decorate/Opacity/Opacity.swift
@@ -7,21 +7,23 @@
 
 import Foundation
 
-/// 색상의 투명도를 정의한 Enum 타입입니다.
-public enum Opacity {
-    case p000
-    case p005
-    case p010
-    case p015
-    case p020
-    case p030
-    case p040
-    case p050
-    case p060
-    case p070
-    case p080
-    case p090
-    case p095
-    case p099
-    case p100
+extension Decorate {
+    /// 색상의 투명도를 정의한 Enum 타입입니다.
+    public enum Opacity {
+        case p000
+        case p005
+        case p010
+        case p015
+        case p020
+        case p030
+        case p040
+        case p050
+        case p060
+        case p070
+        case p080
+        case p090
+        case p095
+        case p099
+        case p100
+    }
 }
